# perf: streaming values up to 50× faster