--- a/.gitignore
+++ b/.gitignore
@@ -3,6 +3,20 @@
 actionlint
 
 /crates/jsonmodem-py/target/
+# jsonmodem-py (Python bindings) build artifacts
+/crates/jsonmodem-py/target/
+/crates/jsonmodem-py/.pytest_cache/
+/crates/jsonmodem-py/python/jsonmodem/_jsonmodem*.so
+/crates/jsonmodem-py/python/jsonmodem/_jsonmodem*.pyd
+/crates/jsonmodem-py/python/jsonmodem/__pycache__/
+
+# General Python caches anywhere in the repo
+__pycache__/
+*.py[cod]
 .venv
 .agent/.setup_py_done
 .agent/.setup_done
+/flamegraph.svg
+/perf.data
+/perf.data.old
+/recent.log