--- a/crates/jsonmodem/benches/streaming_parser.rs
+++ b/crates/jsonmodem/benches/streaming_parser.rs
@@ -28,7 +28,8 @@ fn make_json_payload(target_len: usize) -> String {
     s.push_str("{\"data\":\"");
     s.extend(std::iter::repeat_n('a', content_len));
     s.push_str("\"}");
-    debug_assert_eq!(s.len(), target_len);
+    #[cfg(any(fuzzing, debug_assertions))]
+    assert_eq!(s.len(), target_len);
     s
 }