--- a/crates/jsonmodem/src/value_tree.rs
+++ b/crates/jsonmodem/src/value_tree.rs
@@ -1,0 +1,1 @@
+pub(crate) use crate::backend::value_tree::ValueTree;