… .agent/SPECS.md
… .config/nextest.toml
… .github/workflows/benchmarks.yml
… .github/workflows/fuzz.yml
… .github/workflows/lint.yml
… .github/workflows/miri.yml
… .github/workflows/publish-pypi.yml
… .github/workflows/test.yml
… AGENTS.md
… Cargo.lock
… README.md
… buffer.md
… crates/jsonmodem-py/Cargo.toml
… crates/jsonmodem-py/README.md
… crates/jsonmodem-py/pyproject.toml
… crates/jsonmodem-py/python/jsonmodem/py.typed
… crates/jsonmodem-py/tests/test_import.py
… crates/jsonmodem/.agent/specs/jsonmodem-buffers-rewrite.md
… crates/jsonmodem/Cargo.toml
… crates/jsonmodem/benches/streaming_json_medium.rs
… crates/jsonmodem/examples/llm_tool_call.rs
… crates/jsonmodem/src/backend/raw.rs
… crates/jsonmodem/src/backend/std/mod.rs
… crates/jsonmodem/src/backend/std/value.rs
… crates/jsonmodem/src/backend/std/value_applicator.rs
… crates/jsonmodem/src/backend/std/value_tree.rs
… crates/jsonmodem/src/backend/std/value_zipper.rs
… crates/jsonmodem/src/backend/transition_debug.rs
… crates/jsonmodem/src/backend/zipper_transition.rs
… crates/jsonmodem/src/buffer_options.rs