Performance highlights (from README):

- Streaming values adapter: 5,000 chunks 14,358 μs → 274 μs (~52× faster), 1,000 chunks 3,078 μs → 202 μs (~15.2×), 100 chunks 426 μs → 175 μs (~2.4×).

This commit rebuilds the streaming JSON core to use lending iterators (via a GAT‐based `LendingIterator` trait) and splits higher‑level behavior into small adapters. The result is a leaner, zero‑copy event core with clearer lifetimes and better chunked‑input ergonomics.

Docs

- Fix README examples to use lending‑iterator style:
  - Quick start: iterate with `while let Some(ev) = it.next()`.
  - Buffered strings: call `.to_iter()` when collecting into a `Vec<_>`.

Highlights

- Core now exposes a lending iterator:
  - `JsonModem::feed(&str) -> JsonModemIterator` and `JsonModem::finish() -> JsonModemClosed`
  - Both implement `lending_iterator::LendingIterator` with `Item<'a>` that borrows from the iterator during each `next()` call.
- New adapters on top of the core:
  - `JsonModemBuffers` → coalesces string fragments and can attach full values or prefixes (yields `BufferedEvent`).
  - `JsonModemValues` → incrementally builds partial/complete values (yields `StreamingValue`).
- Parser internals are modularized under `parser/` (scanner, buffers, options, tests).
- New backend layer (`backend::{StdBackend, ...}`) decouples event/value types from the core.
- Crate core is `no_std` with `alloc`; tests and adapters use `std` as needed.

What changed

- Replace monolithic `src/parser.rs` with `src/parser/`:
  - `JsonModem`, `JsonModemIterator`, `JsonModemClosed`, `ParserOptions`, `ParserError`.
  - New `parser::scanner` and focused helpers for escapes, literals, and event building.
- Introduce `lending_iterator::LendingIterator` and `IntoOwned` traits.
- Events reshaped: `event::ParseEvent` now carries a generic path and backend value types and emits string fragments with `is_initial`/`is_final` markers.
- New adapters:
  - `jsonmodem_buffers::{JsonModemBuffers, BufferedEvent, BufferOptions}`
  - `jsonmodem_values::{JsonModemValues, StreamingValue, ValuesOptions}`
- Backend layer (`backend/std/*`, `backend/raw.rs`) implements path/value policies and assembly.
- Path utilities: `Path`, `PathItem`, `path![]` macro.
- Benchmarks and examples updated to the new API and adapters.
- Fuzzing harness updated; helper `split_into_safe_chunks` is now `pub` with an explicit lifetime.
- CI and local tooling:
  - `AGENTS.md` moved to repo root; `.agent/check.sh` updated (fast flags for local runs).
  - `nextest` config added; workflows adjusted; README refreshed.

Why lending iterators

- Zero‑copy where possible: event payloads and paths can borrow input/adapters for each yield.
- Clear lifetime boundaries: each `next()` returns an item bound to the iterator’s borrow.
- Better streaming ergonomics: safe across chunk boundaries without eager buffering.

Breaking changes

- Old APIs removed/renamed:
  - `StreamingParser`, `streaming_values`, `factory.rs`, `event_stack.rs` removed.
  - Former `src/value_zipper.rs` moved under `backend/std/` and is not a public surface.
  - Options reworked; use `parser::ParserOptions` (+ `BufferOptions` / `ValuesOptions` in adapters).
- Event shape changed:
  - Strings are emitted as fragments with `is_initial`/`is_final`; composite values are produced only by adapters.
- Core crate is `no_std` (requires `alloc`).

Migration guide (Rust)

Old (conceptual):

```rust
// let mut p = StreamingParser::new(ParserOptions { /* ... */ });
// for chunk in chunks { p.feed(chunk); }
// for ev in p.finish() { /* handle events */ }
```

New – buffered strings/values via adapter:

```rust
use jsonmodem::{JsonModemBuffers, ParserOptions, BufferOptions, BufferedEvent, path};

let mut modem = JsonModemBuffers::new(ParserOptions::default(), BufferOptions::default());
for chunk in chunks {
    for ev in modem.feed(chunk) {
        match ev.unwrap() {
            BufferedEvent::String { path, fragment, is_final, .. } => { /* ... */ }
            _ => {}
        }
    }
}
for ev in modem.finish() { /* drain remaining */ }
```

Notes

- Parser options include Unicode‐whitespace and multi‑value toggles plus Unicode decode mode controls; lossy/strict behavior is handled by the backend.
- Benchmarks/examples/fuzz updated to reflect the new layers.